+++ src/compiler/entries/entry-modules.ts
@@ -3,7 +3,7 @@ import { buildWarn, catchError } from '../util';
 import { calcComponentDependencies } from './component-dependencies';
 import { DEFAULT_STYLE_MODE, ENCAPSULATION } from '../../util/constants';
 import { generateComponentEntries } from './entry-components';
-import { requiresScopedStyles } from '../style/style';
+import { requiresScopedStyles } from '../style/generate-component-styles';
 import { validateComponentTag } from '../config/validate-component';