+++ test/karma/test-app/attribute-host/karma.spec.ts
@@ -21,17 +21,27 @@ describe('attribute-host', function() {
     expect(elm.getAttribute('color')).toBe(null);
     expect(elm.getAttribute('no-attr')).toBe(null);
 
-    expect(elm.style.getPropertyValue('border-color')).toEqual('black');
+    expect(elm.style.getPropertyValue('border-color')).toEqual('');
+    expect(elm.style.getPropertyValue('display')).toEqual('inline-block');
+    expect(elm.style.getPropertyValue('font-size')).toEqual('');
 
     // this tests CSS custom properties in inline style, but CSS var are
     // not supported natively in IE11, so let's skip the test
     const win = window as any;
     if (win.CSS && win.CSS.supports && win.CSS.supports('--prop', 'value')) {
-      expect(elm.style.getPropertyValue('--css-var')).toEqual('12');
+      expect(elm.style.getPropertyValue('--css-var')).toEqual('');
     }
 
     button.click();
     await flush(app);
+
+    expect(elm.style.getPropertyValue('border-color')).toEqual('black');
+    expect(elm.style.getPropertyValue('display')).toEqual('block');
+    expect(elm.style.getPropertyValue('font-size')).toEqual('24px');
+
+    if (win.CSS && win.CSS.supports && win.CSS.supports('--prop', 'value')) {
+      expect(elm.style.getPropertyValue('--css-var')).toEqual('12');
+    }
 
     expect(elm.getAttribute('content')).toBe('attributes added');
     expect(elm.getAttribute('padding')).toBe('true');
@@ -43,6 +53,14 @@ describe('attribute-host', function() {
     button.click();
     await flush(app);
 
+    expect(elm.style.getPropertyValue('border-color')).toEqual('');
+    expect(elm.style.getPropertyValue('display')).toEqual('inline-block');
+    expect(elm.style.getPropertyValue('font-size')).toEqual('');
+
+    if (win.CSS && win.CSS.supports && win.CSS.supports('--prop', 'value')) {
+      expect(elm.style.getPropertyValue('--css-var')).toEqual('');
+    }
+
     expect(elm.getAttribute('content')).toBe('attributes removed');
     expect(elm.getAttribute('padding')).toBe('false');
     expect(elm.getAttribute('bold')).toBe('false');
@@ -52,6 +70,14 @@ describe('attribute-host', function() {
 
     button.click();
     await flush(app);
+
+    expect(elm.style.getPropertyValue('border-color')).toEqual('black');
+    expect(elm.style.getPropertyValue('display')).toEqual('block');
+    expect(elm.style.getPropertyValue('font-size')).toEqual('24px');
+
+    if (win.CSS && win.CSS.supports && win.CSS.supports('--prop', 'value')) {
+      expect(elm.style.getPropertyValue('--css-var')).toEqual('12');
+    }
 
     expect(elm.getAttribute('content')).toBe('attributes added');
     expect(elm.getAttribute('padding')).toBe('true');